requirements.txt
# requirements.txt
PyQt5==5.15.7
dnspython==2.3.0
